Fix tests and add dep.
