+++ setup.py
@@ -30,7 +30,7 @@ REQUIRED_PACKAGES = [
     "requests",
     "jinja2",
     "networkx",
-    "junit_xml",
+    "junit-xml",
     #"matplotlib", broken on ppc
 ]
